# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tax & Accounting Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalClients: number;
  pendingTasks: number;
  overdueTasks: number;
  completedTasks: number;
  vatDueThisMonth: number;
  corporateTaxDueThisMonth: number;
}